Chia Components cho các Page
