+++ src/routes/index.jsx
@@ -1,14 +1,14 @@
 import { Route, Routes } from "react-router-dom";
 import MainLayout from "../layouts/MainLayout";
-import HomePage from "../pages/HomePage";
-import DetailPage from "../pages/DetailPage";
+import HomePage from "../pages/home-page";
+import DetailPage from "../pages/detail-page/index";
 import BlankLayout from "../layouts/BlankLayout";
 import NotFoundPage from "../pages/NotFoundPage";
 import LoginPage from "@/pages/LoginPage";
 import RegisterPage from "@/pages/RegisterPage";
-import ProductsPage from "@/pages/ProductsPage";
+import ProductsPage from "@/pages/product-page/index";
 import AuthRequired from "./AuthRequired";
-import DashBoardPage from "@/pages/dashboard/DashBoardPage";
+import DashBoardPage from "@/pages/dash-board/DashBoardPage";
 
 function Router() {
   return (
